Fix compiler warning for implicit conversions. Code cleanup

+++ CitySimulator/Controllers/Government.cpp
@@ -2,12 +2,12 @@
 #include "CoreController.h"
 
 
-float Government::taxDollars = 0;
-float Government::birthRate = 0.3;
+float Government::taxDollars = 0.f;
+float Government::birthRate = 0.3f;
 
 int Government::NewChildren(const int adultCount)
 {
-    const int thresh = birthRate * 100;
+    const int thresh = static_cast<int>(birthRate * 100);
     int birthCount = 0;
     for (int i = 0; i < adultCount; ++++i)
     {
